% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello.R
\name{trello}
\alias{trello}
\title{Trello}
\usage{
trello(key, token)
}
\arguments{
\item{key}{string your trello access key}

\item{token}{string your trello access token}
}
\description{
Trello object to store key and token variables
}
\examples{
## create trello object to store your key and token
trello <- trello(key = "abc", token = "xyz")
}

