% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_boards.R
\name{trello_board}
\alias{trello_board}
\title{Trello Board}
\usage{
trello_board(board_id, key, token)
}
\arguments{
\item{board_id}{the board id - found in the URL of a given board - https://trello.com/b/board_id/board_name}

\item{key}{string your trello access key}

\item{token}{string your trello access token}
}
\description{
Retrieve information for a specific board
}

