% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_boards.R
\name{trello_boards}
\alias{trello_boards}
\title{Trello Boards}
\usage{
trello_boards(key, token)
}
\arguments{
\item{key}{string your trello access key}

\item{token}{string your trello access token}
}
\description{
Retreive all boards associated with your project
}

