% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_card.R
\name{trello_card}
\alias{trello_card}
\title{Trello Cards}
\usage{
trello_card(card_id, trello)
}
\arguments{
\item{card_id}{string card id}

\item{trello}{trello object containing your key and token}
}
\description{
Retrieve all cards on a board
}

