% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_list.R
\name{trello_list}
\alias{trello_list}
\title{Trello lists}
\usage{
trello_list(list_id, trello)
}
\arguments{
\item{list_id}{string list id}

\item{trello}{trello object containing your key and token}
}
\description{
Retrieve all cards in a list
}

