% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_organisation.R
\name{trello_organisation}
\alias{trello_organisation}
\title{Trello Organisation}
\usage{
trello_organisation(organisation, trello)
}
\arguments{
\item{organisation}{string name of organisation}

\item{trello}{object containing your key and token}
}
\description{
Retrieve informatin about your organisation on trello
}

