% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trello_boards.R
\name{trello_boards}
\alias{trello_boards}
\title{Trello Boards}
\usage{
trello_boards(trello)
}
\arguments{
\item{trello}{trello object containing your key and token}
}
\description{
Retrieve all boards associated with your project
}

